refactor DefaultSecurityConfig location
