+++ gatekeeper/src/main/java/eu/arrowhead/core/gatekeeper/security/GKSecurityConfig.java
@@ -3,7 +3,7 @@ package eu.arrowhead.core.gatekeeper.security;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
 
-import eu.arrowhead.common.DefaultSecurityConfig;
+import eu.arrowhead.common.security.DefaultSecurityConfig;
 
 @Configuration
 @EnableWebSecurity
